counting holes doesn't assume 4-hour timetable
 - split Fitness into Fitness and FitnessCalculator so that
FitnessCalculator is stateless
 - new algorithm for counting holes in the timetable which does not
assume 4-hour timetables
 - new algorithm which uses the logistic function to translate the
number of holes into a penalty. The constants were chosen in such a way
that the new algorithm gives the same results as the old one for
a 4-hour timetable. (This was achieved by solving the series of
equations 0.2=1/(1+exp(-k*(1/4-x))) and 0.8=1/(1+exp(-k*(1/2-x))) using
Wolfram Alpha. 0.2 corresponds to penalty1UnwantedBetweenWanted and 0.8
corresponds to penalty1UnwantedBetweenWanted
+ penalty2UnwantedBetweenWanted. 1/4 and 1/2 are the hole counts
corresponding to these penalties.)